Refactorizo y limpio el codigo, intercambio de clave y mensajeria encriptada completamente funcional
